source.requirements.txt
--extra-index-url https://karrio.gateway.scarf.sh/simple/
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
cssselect2==0.8.0
deprecation==2.1.0
diff-match-patch==20241021
distro==1.9.0
dj-database-url==3.0.0
Django==5.2.3
django-constance==4.3.2
django-cors-headers==4.7.0
django-downloadview==2.4.0
django-email-verification==0.3.3
django-filter==25.1
django-formtools==2.5.1
django-health-check==3.20.0
django-import-export==4.3.8
django-oauth-toolkit==3.0.1
django-otp==1.6.0
django-phonenumber-field==8.1.0
django-picklefield==3.3
django-redis==6.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-api-tracking==1.8.4
drf-spectacular==0.28.0
fonttools==4.58.4
graphql-core==3.2.6
gunicorn==23.0.0
h11==0.16.0
hiredis==3.2.1
huey==2.5.3
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jsonfield==3.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jstruct==2021.11
jwcrypto==1.5.6
karrio @ file://${PWD}/modules/sdk
karrio_australiapost @ file://${PWD}/modules/connectors/australiapost
karrio_canadapost @ file://${PWD}/modules/connectors/canadapost
karrio_chronopost @ file://${PWD}/community/plugins/chronopost
karrio_dhl_express @ file://${PWD}/modules/connectors/dhl_express
karrio_dhl_parcel_de @ file://${PWD}/modules/connectors/dhl_parcel_de
karrio_dhl_poland @ file://${PWD}/modules/connectors/dhl_poland
karrio_dhl_universal @ file://${PWD}/modules/connectors/dhl_universal
karrio_dpd @ file://${PWD}/community/plugins/dpd
karrio_easypost @ file://${PWD}/community/plugins/easypost
karrio_easyship @ file://${PWD}/community/plugins/easyship
karrio_eshipper @ file://${PWD}/community/plugins/eshipper
karrio_fedex @ file://${PWD}/modules/connectors/fedex
karrio_freightcom @ file://${PWD}/community/plugins/freightcom
karrio_generic @ file://${PWD}/modules/connectors/generic
karrio_laposte @ file://${PWD}/modules/connectors/laposte
karrio_purolator @ file://${PWD}/modules/connectors/purolator
karrio_sapient @ file://${PWD}/community/plugins/sapient
karrio_seko @ file://${PWD}/modules/connectors/seko
karrio_sendle @ file://${PWD}/modules/connectors/sendle
karrio_server @ file://${PWD}/apps/api
karrio_server_core @ file://${PWD}/modules/core
karrio_server_data @ file://${PWD}/modules/data
karrio_server_documents @ file://${PWD}/modules/documents
karrio_server_events @ file://${PWD}/modules/events
karrio_server_graph @ file://${PWD}/modules/graph
karrio_server_manager @ file://${PWD}/modules/manager
karrio_server_orders @ file://${PWD}/modules/orders
karrio_server_pricing @ file://${PWD}/modules/pricing
karrio_server_proxy @ file://${PWD}/modules/proxy
karrio_ups @ file://${PWD}/modules/connectors/ups
karrio_usps @ file://${PWD}/modules/connectors/usps
karrio_usps_international @ file://${PWD}/modules/connectors/usps_international
lxml==4.9.4
lxml-stubs==0.5.1
MarkupSafe==3.0.2
more-itertools==10.7.0
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.8
pillow==11.2.1
posthog==5.4.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-soap==2020.3.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
PyPDF2==3.0.1
pyphen==0.17.2
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
pyzint==0.1.10
qrcode==7.4.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
sentry-sdk==2.31.0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.275.5
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.35.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
xmltodict==0.14.2
zopfli==0.2.3.post1

karrio_logistic_io @ file://${PWD}/modules/connectors/logistic_io
karrio_aymakan @ file://${PWD}/modules/connectors/aymakan
karrio_jtexpress @ file://${PWD}/modules/connectors/jtexpress
